separate keywords functionality from grants store
