+++ packages/client/src/store/index.js
@@ -9,6 +9,7 @@ import dashboard from './modules/dashboard';
 import organization from './modules/organization';
 import tenants from './modules/tenants';
 import alerts from './modules/alerts';
+import keywords from './modules/keywords';
 
 Vue.use(Vuex);
 
@@ -25,5 +26,6 @@ export default new Vuex.Store({
     organization,
     tenants,
     alerts,
+    keywords,
   },
 });
